third_party/WebKit/LayoutTests/fast/events/touch/compositor-touch-hit-rects-iframes-expected.txt

This tests verifies the hit test regions given to the compositor when non-composited iframes are involved. It can only be run in DumpRenderTree. The outputted rects should cover the hit test regions of all the listed elements. Enable visualize mode to quickly validate graphically.

iframe: #document scrolling (23, 123, 290, 22)

iframe-nested: #document scrolling (33, 161, 290, 22)

iframe-transform: #document scrolling (34, 199, 320, 31)

iframe-fixed: #document scrolling (31, 262, 332, 22)

iframe-doc: #document scrolling (10, 284, 400, 25)
